+++ src/main/java/tconstruct/mechworks/entity/item/ExplosivePrimed.java
@@ -1,36 +1,40 @@
 package tconstruct.mechworks.entity.item;
 
-import cpw.mods.fml.relauncher.*;
 import net.minecraft.block.Block;
-import net.minecraft.entity.*;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.world.*;
+import net.minecraft.world.Explosion;
+import net.minecraft.world.World;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import tconstruct.world.MiningExplosion;
 
-public class ExplosivePrimed extends Entity
-{
+public class ExplosivePrimed extends Entity {
+
     /** How long the fuse is */
     public Block block;
+
     public int metadata;
     public int fuse;
     private EntityLivingBase tntPlacedBy;
 
-    public ExplosivePrimed(World par1World)
-    {
+    public ExplosivePrimed(World par1World) {
         super(par1World);
         this.preventEntitySpawning = true;
         this.setSize(0.98F, 0.98F);
         this.yOffset = this.height / 2.0F;
     }
 
-    public ExplosivePrimed(World par1World, double par2, double par4, double par6, EntityLivingBase par8EntityLivingBase)
-    {
+    public ExplosivePrimed(World par1World, double par2, double par4, double par6,
+            EntityLivingBase par8EntityLivingBase) {
         this(par1World);
         this.setPosition(par2, par4, par6);
         float f = (float) (Math.random() * Math.PI * 2.0D);
-        this.motionX = (double) (-((float) Math.sin((double) f)) * 0.02F);
+        this.motionX = -((float) Math.sin(f)) * 0.02F;
         this.motionY = 0.20000000298023224D;
-        this.motionZ = (double) (-((float) Math.cos((double) f)) * 0.02F);
+        this.motionZ = -((float) Math.cos(f)) * 0.02F;
         this.fuse = 80;
         this.prevPosX = par2;
         this.prevPosY = par4;
@@ -39,27 +43,22 @@ public class ExplosivePrimed extends Entity
     }
 
     @Override
-    protected void entityInit ()
-    {
-    }
+    protected void entityInit() {}
 
     /**
-     * returns if this entity triggers Block.onEntityWalking on the blocks they
-     * walk on. used for spiders and wolves to prevent them from trampling crops
+     * returns if this entity triggers Block.onEntityWalking on the blocks they walk on. used for spiders and wolves to
+     * prevent them from trampling crops
      */
     @Override
-    protected boolean canTriggerWalking ()
-    {
+    protected boolean canTriggerWalking() {
         return false;
     }
 
     /**
-     * Returns true if other Entities should be prevented from moving through
-     * this Entity.
+     * Returns true if other Entities should be prevented from moving through this Entity.
      */
     @Override
-    public boolean canBeCollidedWith ()
-    {
+    public boolean canBeCollidedWith() {
         return !this.isDead;
     }
 
@@ -67,8 +66,7 @@ public class ExplosivePrimed extends Entity
      * Called to update the entity's position/logic.
      */
     @Override
-    public void onUpdate ()
-    {
+    public void onUpdate() {
         this.prevPosX = this.posX;
         this.prevPosY = this.posY;
         this.prevPosZ = this.posZ;
@@ -78,36 +76,29 @@ public class ExplosivePrimed extends Entity
         this.motionY *= 0.9800000190734863D;
         this.motionZ *= 0.9800000190734863D;
 
-        if (this.onGround)
-        {
+        if (this.onGround) {
             this.motionX *= 0.699999988079071D;
             this.motionZ *= 0.699999988079071D;
             this.motionY *= -0.5D;
         }
 
-        if (this.fuse-- <= 0)
-        {
+        if (this.fuse-- <= 0) {
             this.setDead();
 
-            if (!this.worldObj.isRemote)
-            {
+            if (!this.worldObj.isRemote) {
                 this.explode();
             }
-        }
-        else
-        {
+        } else {
             this.worldObj.spawnParticle("smoke", this.posX, this.posY + 0.5D, this.posZ, 0.0D, 0.0D, 0.0D);
         }
     }
 
-    private void explode ()
-    {
+    private void explode() {
         float f = 5.0F;
         this.createExplosion(worldObj, this, this.posX, this.posY, this.posZ, f);
     }
 
-    public void createExplosion (World world, Entity par1Entity, double par2, double par4, double par6, float par8)
-    {
+    public void createExplosion(World world, Entity par1Entity, double par2, double par4, double par6, float par8) {
         Explosion explosion = new MiningExplosion(world, par1Entity, par2, par4, par6, par8);
         explosion.isFlaming = false;
         explosion.isSmoking = true;
@@ -120,8 +111,7 @@ public class ExplosivePrimed extends Entity
      * (abstract) Protected helper method to write subclass entity data to NBT.
      */
     @Override
-    protected void writeEntityToNBT (NBTTagCompound par1NBTTagCompound)
-    {
+    protected void writeEntityToNBT(NBTTagCompound par1NBTTagCompound) {
         par1NBTTagCompound.setByte("Fuse", (byte) this.fuse);
     }
 
@@ -129,23 +119,20 @@ public class ExplosivePrimed extends Entity
      * (abstract) Protected helper method to read subclass entity data from NBT.
      */
     @Override
-    protected void readEntityFromNBT (NBTTagCompound par1NBTTagCompound)
-    {
+    protected void readEntityFromNBT(NBTTagCompound par1NBTTagCompound) {
         this.fuse = par1NBTTagCompound.getByte("Fuse");
     }
 
     @Override
     @SideOnly(Side.CLIENT)
-    public float getShadowSize ()
-    {
+    public float getShadowSize() {
         return 0.0F;
     }
 
     /**
      * returns null or the entityliving it was placed or ignited by
      */
-    public EntityLivingBase getTntPlacedBy ()
-    {
+    public EntityLivingBase getTntPlacedBy() {
         return this.tntPlacedBy;
     }
 }
